Fix getProperty recursive call
When getProperty called and in the callback the getProperty is called
again with the same property, the second callback would never be called
because it would be cleaned from the getter array.

viewer.getProperty('playingContent', function () {
	console.log('Callback 1', content);

	viewer.getProperty('playingContent', function () {
		console.log('Callback 2', content);
	});
});

The Callback 2 never been written.

Fix the bug by keeping a reference to the getter array and remove the
reference to the array from the `getters` object. In that case a new
getProperty call initializes a new array for the event.